+++ pubspec.yaml
@@ -54,6 +54,7 @@ dependencies:
   package_info_plus: ^5.0.1
   install_plugin: ^2.1.0
   permission_handler: ^11.3.1
+  flustars: ^2.0.1
 
 dev_dependencies:
   flutter_test:
